+++ packages/react-microlink/src/components/Card/CardMedia/index.js
@@ -12,32 +12,28 @@ export default class CardMedia extends Component {
   constructor (props) {
     super(props)
     this.state = {
-      loadingImage: isUrl(props.image)
+      loading: isUrl(props.image)
     }
   }
 
-  imageLoaded () {
-    this.setState({loadingImage: false})
-  }
-
   renderMedia () {
-    const {loadingImage} = this.state
+    const {loading} = this.state
     const {image, video} = this.props
     const el = !isUrl(video) && isUrl(image) ? Image : Video
     return createElement(el, {
       ...this.props,
       key: 'media',
-      loading: loadingImage
+      loading
     })
   }
 
   renderLoadCatcher () {
     const {image} = this.props
-    const {loadingImage: loading} = this.state
+    const {loading} = this.state
     return loading && isUrl(image) && createElement(ImageLoadCatcher, {
       key: 'imageLoader',
       src: image,
-      onLoad: () => this.imageLoaded()
+      onLoad: () => this.setState({loading: false})
     })
   }