+++ packages/react-microlink/src/components/Card/CardWrap.js
@@ -7,9 +7,12 @@ const HEIGHT = '382px'
 
 const contrastStyle = ({backgroundColor, color}) => css`
   background-color: ${backgroundColor};
-  color: ${color};
   border-color: ${color};
-  transition: filter .15s ease-in-out;
+  transition-property: filter;
+
+  &&& {
+    color: ${color};
+  }
 
   &:hover {
     filter: brightness(90%);
@@ -19,15 +22,15 @@ const contrastStyle = ({backgroundColor, color}) => css`
 const largeStyle = css`
   flex-direction: column;
   height: ${HEIGHT};
-  transition: height .25s ease-in-out;
+  transition-property: background, border-color, height;
 
   ${media.mobile`
     height: calc(${HEIGHT} * 7/9);
   `}
 `
 
-const loadingStyle = css`
-  transition: background-color .15s ease-in-out, border-color .15s ease-in-out;
+const hoverStyle = css`
+  transition-property: background, border-color;
   &:hover {
     background: #F5F8FA;
     border-color: rgba(136,153,166,.5);
@@ -39,7 +42,7 @@ const reverseStyle = ({cardSize}) => css`
 `
 
 const roundStyle = ({round}) => css`
-  border-radius: ${typeof round === 'boolean' ? `.42857em` : round};
+  border-radius: ${round === true ? `.42857em` : round};
 `
 
 const style = css`
@@ -49,17 +52,22 @@ const style = css`
   border-style: solid;
   border-color: #E1E8ED;
   overflow: hidden;
+  color: #181919;
+  font-family: 'Helvetica Neue', Helvetica, Arial, sans-serif;
   display: flex;
   text-decoration: none;
   opacity:1;
   position: relative;
+
+  transition-duration: .15s;
+  transition-timing-function: ease-in-out;
 
   &:active,
   &:hover {
     outline: 0;
   }
 
-  ${({loading, contrast}) => !loading && !contrast && loadingStyle}
+  ${({loading, contrast}) => !loading && !contrast && hoverStyle}
 
   ${({round}) => round && roundStyle}
 
@@ -68,6 +76,8 @@ const style = css`
   ${({reverse}) => reverse && reverseStyle}
 
   ${({backgroundColor, color, contrast}) => contrast && color && backgroundColor && contrastStyle}
+
+  ${({backgroundColor, color, contrast}) => contrast && (!color || !backgroundColor) && hoverStyle}
 `
 
 const CardWrap = ({ is, rel, href, target, ...props }) => {